updated ft_strlen() to be of type size_t instead of unsigned int.

+++ ft_strlen.c
@@ -6,16 +6,18 @@
 /*   By: okuilboe <okuilboe@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/04/25 12:45:19 by okuilboe      #+#    #+#                 */
-/*   Updated: 2025/04/28 17:33:41 by okuilboe      ########   odam.nl         */
+/*   Updated: 2025/05/01 20:51:28 by okuilboe      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
 		i++;
-	return (i);
+	return ((size_t)i);
 }
